Revert "more field->channel renaming"
This reverts commit 87b79906696407131345263968f11d184310c98a.
