+++ frontends/api/src/test-utils/factories/index.ts
@@ -2,4 +2,4 @@ export * as learningResources from "./learningResources"
 export * as userLists from "./userLists"
 export * as articles from "./articles"
 export * as letters from "./programLetters"
-export * as fields from "./channels"
+export * as fields from "./fields"
