+++ frontends/mit-open/src/pages/ChannelPage/ChannelPage.tsx
@@ -3,7 +3,7 @@ import { useParams, useLocation, useNavigate } from "react-router"
 import { Tab, TabContext, TabList, TabPanel, Container } from "ol-components"
 import { Link } from "react-router-dom"
 import FieldPageSkeleton from "./ChannelPageSkeleton"
-import { useChannelDetailByType } from "../../../../api/src/hooks/channels"
+import { useChannelDetailByType } from "api/hooks/fields"
 import WidgetsList from "./WidgetsList"
 import { GridColumn, GridContainer } from "@/components/GridLayout/GridLayout"
 import { makeFieldViewPath } from "@/common/urls"
